feat: finish prueba03
