+++ pruebas/03/src/models/task.model.tsx
@@ -1,0 +1,4 @@
+export interface TaskModel {
+  id: number;
+  name: string;
+}
